Always display the selected language
The selected language was not displayed when it was not among the
languages proposed in the menu. We now add a line to the menu for the
selected language if not initially present in the menu. This line gets
updated each time such a language is selected, so the menu doens't grow
too much.