Put chosen layout first in /etc/default/keyboard.
This makes e.g. "es,us" instead of "us,es" written into there, so that in turn
the console keyboard layout is properly localized instead of being set to
US qwerty.

Rationale for not reversing the layout list in the langpanel class:
xklavier needs the list in "last item prevails" order.