Apply setting changes immediately and not in a deleter
Relying on deleters to perform critical tasks, in that case, providing the
data needed for the rest of the session configuration process, should not be
done in deleters. We have no guarantees that they will be called in case of
signals or exits.

As we are all working in RAM, writing a short configuration snippet should not
take long. Let's just do that synchronously everytime a setting is changed,
that will help to detect bugs and prevent races from happening.