Added new file "default_langcodes"
This file lists the default language country pairs,
e.g. en_US or sv_SE. This is needed because the latter
is currently incorrectly set to sv_FI when sv is selected,
for example.

The list was derived from a similar file in Ubiquity,
which in turn appears to have derived it from
Debian Installer's localechooser.