Fix link activation via Return key press.
Previously it was also caught by our custom login shortcut, which also
uses Return as the trigger key.

As a side note it can be said that links inside labels seem impossible
to make unfocusable, even if the labels are configured to be so. In
other words, the introduction of the links will break any TAB or arrow
key-based login procedure users may have memorised.