Simply set chosen keyboard layout with no 'us' switch.
The ordering of `layout_list` seem to matter for
Xkl.ConfigRec.set_layouts, but I cannot understand how. It seems
possible to get one layout in the Greeter, and another in GNOME,
depending on their position.

Let's just simplify this so we only set the chosen layout in the
Greeter, and add the 'us' switch in `tails-configure-keyboard`
instead.