Do not reload CUPS configuration in PostLogin.default.
Actually, `service cups force-reload' forcibly starts the service, even if
we have disabled  cups.service to make CUPS start via socket activation only
(cups.socket). That's not desirable.

Also, it is not necessary anymore: with socket activation, the daemon should not
be running at that point yet, so there's no point in asking it to reload
its configuration.