Import udev trigger + settle from tails-unblock-network.service, and start it in a blocking way.
Apparently, manually messing with udevadm from a systemd unit is racy. It might
be that the races I've seen were actually caused by starting
tails-unblock-network.service in a non-blocking way, but I've spent too much
time on it already, and simply extracting the "trigger kernel events again and
wait for udev to settle" seems to do the trick.