Start tails-unblock-network in the background.
All operations that need to be done in a specific order are now in that
script, so we should be able to avoid some delay in the login process.

This is quite important for UX since this delay happens at a time when
we provide little info to the user wrt. what's going on.